+++ domains/src/profile/transaction_delete_profile.rs
@@ -1,4 +1,4 @@
-use crate::transaction::Transaction;
+use crate::{transaction::Transaction, errors::{ErrorDomain, profile::ErrorProfile}};
 use repositories::profile_transaction_repository::ProfileTransactionRepository;
 
 pub struct TransactionDeleteProfile<'a> {
@@ -16,8 +16,15 @@ impl TransactionDeleteProfile<'_> {
 }
 
 impl Transaction for TransactionDeleteProfile<'_> {
-    fn execute(&mut self) -> () {
-        self.db.delete_profile(self.profile_id)
+    fn execute(&mut self) ->  Result<(), Box<dyn ErrorDomain>> {
+        let profile = self.db.get_profile(self.profile_id);
+        
+        if !profile.is_none() {
+            self.db.delete_profile(self.profile_id);
+            Ok(())
+        } else {
+            Err(Box::new(ErrorProfile::ProfileNotExist))
+        }
     }
 }
