+++ domains/src/resume/transaction_delete_resume.rs
@@ -1,5 +1,8 @@
 use repositories::resume_transaction_repository::ResumeTransactionRepository;
-use crate::transaction::Transaction;
+use crate::{
+    transaction::Transaction, 
+    errors::{ErrorDomain, resume::ErrorResume}
+};
 
 pub struct TransactionDeleteResume<'a> {
     db: Box<dyn ResumeTransactionRepository + 'a>,
@@ -15,7 +18,14 @@ impl TransactionDeleteResume<'_> {
     }
 }
 impl Transaction for TransactionDeleteResume<'_> {
-    fn execute(&mut self) -> () {
-        self.db.delete_resume(self.resume_id);
+    fn execute(&mut self) -> Result<(), Box<dyn ErrorDomain>> {
+        let resume = self.db.get_resume(self.resume_id);
+
+        if !resume.is_none() {
+            self.db.delete_resume(self.resume_id);
+            Ok(())
+        } else {
+            Err(Box::new(ErrorResume::ResumeNotExist))
+        }
     }
 }
